+++ MaterialColorUtilities/Quantize/IPointProvider.cs
@@ -2,7 +2,7 @@
 
 public interface IPointProvider
 {
-    public float[] FromInt(int argb);
-    public int ToInt(float[] point);
-    public float Distance(float[] a, float[] b);
+    public double[] FromInt(int argb);
+    public int ToInt(double[] point);
+    public double Distance(double[] a, double[] b);
 }
